sem_2_lab_10

+++ 23216/a.kharitonov1/sem_2_lab_10/sem_2_lab_10.c
@@ -13,7 +13,7 @@ pthread_t phils[PHILO];
 void *philosopher (void *id);
 int food_on_table ();
 void get_fork (long long, long long, char *);
-void down_forks (long , long long);
+void down_forks (long long, long long);
 pthread_mutex_t foodlock;
 
 int sleep_seconds = 0;
